Move image utils to lib.image + docstrings
