+++ docs/full/lib.image.rst
@@ -1,0 +1,7 @@
+lib.image module
+========================
+
+.. automodule:: lib.image
+   :members:
+   :undoc-members:
+   :show-inheritance:
