+++ docs/index.rst
@@ -7,7 +7,7 @@ faceswap.dev Developer Documentation
 ====================================
 
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 2
    :caption: Contents:
 
    full/modules
