+++ lib/faces_detect.py
@@ -2,6 +2,7 @@
 """ Face and landmarks detection for faceswap.py """
 import logging
 
+import cv2
 import numpy as np
 
 from lib.aligner import Extract as AlignerExtract, get_align_mat, get_matrix_scaling
@@ -404,3 +405,90 @@ class DetectedFace():
         if not self.reference:
             return None
         return get_matrix_scaling(self.reference_matrix)
+
+
+def rotate_landmarks(face, rotation_matrix):
+    # pylint:disable=c-extension-no-member
+    """ Rotates the 68 point landmarks and detection bounding box around the given rotation matrix.
+
+    Paramaters
+    ----------
+    face: DetectedFace or dict
+        A :class:`DetectedFace` or an `alignments file` ``dict`` containing the 68 point landmarks
+        and the `x`, `w`, `y`, `h` detection bounding box points.
+    rotation_matrix: numpy.ndarray
+        The rotation matrix to rotate the given object by.
+
+    Returns
+    -------
+    DetectedFace or dict
+        The rotated :class:`DetectedFace` or `alignments file` ``dict`` with the landmarks and
+        detection bounding box points rotated by the given matrix. The return type is the same as
+        the input type for ``face``
+    """
+    logger.trace("Rotating landmarks: (rotation_matrix: %s, type(face): %s",
+                 rotation_matrix, type(face))
+    rotated_landmarks = None
+    # Detected Face Object
+    if isinstance(face, DetectedFace):
+        bounding_box = [[face.x, face.y],
+                        [face.x + face.w, face.y],
+                        [face.x + face.w, face.y + face.h],
+                        [face.x, face.y + face.h]]
+        landmarks = face.landmarks_xy
+
+    # Alignments Dict
+    elif isinstance(face, dict) and "x" in face:
+        bounding_box = [[face.get("x", 0), face.get("y", 0)],
+                        [face.get("x", 0) + face.get("w", 0),
+                         face.get("y", 0)],
+                        [face.get("x", 0) + face.get("w", 0),
+                         face.get("y", 0) + face.get("h", 0)],
+                        [face.get("x", 0),
+                         face.get("y", 0) + face.get("h", 0)]]
+        landmarks = face.get("landmarks_xy", list())
+
+    else:
+        raise ValueError("Unsupported face type")
+
+    logger.trace("Original landmarks: %s", landmarks)
+
+    rotation_matrix = cv2.invertAffineTransform(  # pylint:disable=no-member
+        rotation_matrix)
+    rotated = list()
+    for item in (bounding_box, landmarks):
+        if not item:
+            continue
+        points = np.array(item, np.int32)
+        points = np.expand_dims(points, axis=0)
+        transformed = cv2.transform(points,  # pylint:disable=no-member
+                                    rotation_matrix).astype(np.int32)
+        rotated.append(transformed.squeeze())
+
+    # Bounding box should follow x, y planes, so get min/max
+    # for non-90 degree rotations
+    pt_x = min([pnt[0] for pnt in rotated[0]])
+    pt_y = min([pnt[1] for pnt in rotated[0]])
+    pt_x1 = max([pnt[0] for pnt in rotated[0]])
+    pt_y1 = max([pnt[1] for pnt in rotated[0]])
+    width = pt_x1 - pt_x
+    height = pt_y1 - pt_y
+
+    if isinstance(face, DetectedFace):
+        face.x = int(pt_x)
+        face.y = int(pt_y)
+        face.w = int(width)
+        face.h = int(height)
+        face.r = 0
+        if len(rotated) > 1:
+            rotated_landmarks = [tuple(point) for point in rotated[1].tolist()]
+            face.landmarks_xy = rotated_landmarks
+    else:
+        face["left"] = int(pt_x)
+        face["top"] = int(pt_y)
+        face["right"] = int(pt_x1)
+        face["bottom"] = int(pt_y1)
+        rotated_landmarks = face
+
+    logger.trace("Rotated landmarks: %s", rotated_landmarks)
+    return face
